push shared artifacts up to new abstract base class
